Refactored Selection Sort: fixed filename, added dynamic memory, optimization, and docs
