+++ NOTES/Bubble.md
